+++ lib/onlyoffice_webdriver_wrapper/webdriver/click_methods.rb
@@ -3,21 +3,16 @@
 module OnlyofficeWebdriverWrapper
   # Method to perform different clicks
   module ClickMethods
+    # Click on specified element
+    # @param element [Selenium::WebDriver::Element] element to click
+    # @return [nil] nothing
     def click(element)
       element.click
     end
 
-    def click_and_wait(element_to_click, element_to_wait)
-      element_to_click.click
-      count = 0
-      while !element_to_wait.present? && count < 30
-        sleep 1
-        count += 1
-      end
-      webdriver_error("click_and_wait: Wait for element: #{element_to_click} for 30 seconds") if count == 30
-    end
-
     # Click on locator
+    # @param xpath_name [String] xpath to click
+    # @param by_javascript [True, False] should be clicked by javascript
     # @param count [Integer] count of clicks
     def click_on_locator(xpath_name, by_javascript = false, count: 1)
       element = get_element(xpath_name)
@@ -28,23 +23,31 @@ module OnlyofficeWebdriverWrapper
       else
         begin
           count.times { element.click }
-        rescue Selenium::WebDriver::Error::ElementNotVisibleError
-          webdriver_error("Selenium::WebDriver::Error::ElementNotVisibleError: element not visible for xpath: #{xpath_name}")
+        rescue Selenium::WebDriver::Error::ElementNotVisibleError => e
+          webdriver_error(e.class, "Selenium::WebDriver::Error::ElementNotVisibleError: element not visible for xpath: #{xpath_name}")
         rescue Exception => e
           webdriver_error(e.class, "UnknownError #{e.message} #{xpath_name}")
         end
       end
     end
 
+    # Click on one of several which displayed
+    # @param xpath_name [String] xpath to find element
+    # @return [nil]
     def click_on_displayed(xpath_name)
       element = get_element_by_display(xpath_name)
       begin
         element.is_a?(Array) ? element.first.click : element.click
       rescue Exception => e
-        webdriver_error("Exception #{e} in click_on_displayed(#{xpath_name})")
+        webdriver_error(e.class, "Exception #{e} in click_on_displayed(#{xpath_name})")
       end
     end
 
+    # Click on locator by coordinates
+    # @param xpath_name [String] xpath to click
+    # @param right_by [Integer] shift to right
+    # @param down_by [Integer] shift to bottom
+    # @return [nil]
     def click_on_locator_coordinates(xpath_name, right_by, down_by)
       wait_until_element_visible(xpath_name)
       element = @driver.find_element(:xpath, xpath_name)
@@ -52,20 +55,9 @@ module OnlyofficeWebdriverWrapper
       @driver.action.move_to(element, right_by.to_i, down_by.to_i).click.perform
     end
 
-    def click_on_one_of_several_by_text(xpath_several_elements, text_to_click)
-      @driver.find_elements(:xpath, xpath_several_elements).each do |current_element|
-        next unless text_to_click.to_s == current_element.attribute('innerHTML')
-
-        begin
-          current_element.click
-        rescue Exception => e
-          webdriver_error("Error in click_on_one_of_several_by_text(#{xpath_several_elements}, #{text_to_click}): #{e.message}")
-        end
-        return true
-      end
-      false
-    end
-
+    # Click on one of several which displayed
+    # @param xpath_several_elements [String] xpath to find element
+    # @return [True, False] true if click successful, false if not found
     def click_on_one_of_several_by_display(xpath_several_elements)
       @driver.find_elements(:xpath, xpath_several_elements).each do |current_element|
         if current_element.displayed?
@@ -76,6 +68,11 @@ module OnlyofficeWebdriverWrapper
       false
     end
 
+    # Click on one of several xpath filtered by parameter and value
+    # @param xpath_several_elements [String] xpath to select
+    # @param parameter_name [String] parameter name
+    # @param parameter_value [String] parameter value
+    # @return [True, False] true if click successful, false if not found
     def click_on_one_of_several_by_parameter(xpath_several_elements, parameter_name, parameter_value)
       @driver.find_elements(:xpath, xpath_several_elements).each do |current_element|
         if current_element.attribute(parameter_name).include? parameter_value
@@ -86,14 +83,20 @@ module OnlyofficeWebdriverWrapper
       false
     end
 
-    def click_on_one_of_several_xpath_by_number(xpath, number_of_element)
-      click_on_locator("(#{xpath})[#{number_of_element}]")
-    end
-
+    # Perform right click on xpath
+    # @param xpath_name [String] xpath to click
+    # @return [nil]
     def right_click(xpath_name)
+      wait_until_element_visible(xpath_name)
+
       @driver.action.context_click(@driver.find_element(:xpath, xpath_name)).perform
     end
 
+    # Perform right click on locator with specified coordinates
+    # @param xpath_name [String] xpath to click
+    # @param right_by [Integer] shift to right
+    # @param down_by [Integer] shift to bottom
+    # @return [nil]
     def right_click_on_locator_coordinates(xpath_name, right_by = nil, down_by = nil)
       wait_until_element_visible(xpath_name)
       element = @driver.find_element(:xpath, xpath_name)
@@ -101,27 +104,22 @@ module OnlyofficeWebdriverWrapper
       @driver.action.move_to(element, right_by.to_i, down_by.to_i).context_click.perform
     end
 
+    # Perform double_click on element
+    # @param xpath_name [String] xpath to click
+    # @return [nil]
     def double_click(xpath_name)
       wait_until_element_visible(xpath_name)
       @driver.action.move_to(@driver.find_element(:xpath, xpath_name)).double_click.perform
     end
 
+    # Perform double_click on specified coordinates
+    # @param xpath_name [String] xpath to click
+    # @param right_by [Integer] shift to right
+    # @param down_by [Integer] shift to bottom
+    # @return [nil]
     def double_click_on_locator_coordinates(xpath_name, right_by, down_by)
       wait_until_element_visible(xpath_name)
       @driver.action.move_to(@driver.find_element(:xpath, xpath_name), right_by.to_i, down_by.to_i).double_click.perform
-    end
-
-    def left_mouse_click(xpath, x_coord, y_coord)
-      @driver.action.move_to(get_element(xpath), x_coord.to_i, y_coord.to_i).click.perform
-    end
-
-    # Context click on locator
-    # @param [String] xpath_name name of xpath to click
-    def context_click_on_locator(xpath_name)
-      wait_until_element_visible(xpath_name)
-
-      element = @driver.find_element(:xpath, xpath_name)
-      @driver.action.context_click(element).perform
     end
   end
 end